Moved `getActivities` action to `api/actions/getActivities`
no refs

Moved the `getActivities` action to `api/actions/getActivities` and updated all
references to it
