+++ src/api/index.ts
@@ -1,3 +1,4 @@
+export { getActivitiesAction } from './action/getActivities';
 export { searchAction } from './action/search';
 export { profileGetFollowersAction } from './action/profile/getFollowers';
 export { profileGetFollowingAction } from './action/profile/getFollowing';
