+++ src/handlers.ts
@@ -2,7 +2,6 @@ import {
     Article,
     Context as APContext,
     Follow,
-    KvStore,
     Like,
     Undo,
     RequestContext,
@@ -14,11 +13,8 @@ import {
     PUBLIC_COLLECTION,
     Mention,
 } from '@fedify/fedify';
-import { Buffer } from 'node:buffer';
 import { Context, Next } from 'hono';
-import sanitizeHtml from 'sanitize-html';
 import { v4 as uuidv4 } from 'uuid';
-import { getActivityMeta, getRepliesMap } from './db';
 import { addToList, removeFromList } from './kv-helpers';
 import { toURL } from './toURL';
 import { ContextData, HonoContextVariables, fedify } from './app';
@@ -28,16 +24,7 @@ import { ACTOR_DEFAULT_HANDLE } from './constants';
 import { Temporal } from '@js-temporal/polyfill';
 import { createHash } from 'node:crypto';
 import { lookupActor } from 'lookup-helpers';
-
-type InboxItem = {
-    id: string;
-    object: string | {
-        id: string;
-        content: string;
-        [key: string]: any;
-    };
-    [key: string]: any;
-}
+import { buildActivity } from 'helpers/activitypub/activity';
 
 import z from 'zod';
 
@@ -486,119 +473,6 @@ export async function siteChangedWebhook(
     });
 }
 
-async function buildActivity(
-    uri: string,
-    db: KvStore,
-    apCtx: APContext<ContextData>,
-    liked: string[] = [],
-    repliesMap: Map<string, any> | null = null,
-    expandInReplyTo: boolean = false,
-): Promise<InboxItem | null> {
-    const item = await db.get<InboxItem>([uri]);
-
-    // If the item is not in the db, return null as we can't build it
-    if (!item) {
-        return null;
-    }
-
-    // If the object associated with the item is a string, it's probably a URI,
-    // so we should look it up in the db. If it's not in the db, we should just
-    // leave it as is
-    if (typeof item.object === 'string') {
-        item.object = await db.get([item.object]) ?? item.object;
-    }
-
-    if (typeof item.actor === 'string') {
-        const actor = await lookupActor(apCtx, item.actor);
-
-        if (actor) {
-            const json = await actor.toJsonLd();
-            if (typeof json === 'object' && json !== null) {
-                item.actor = json;
-            }
-        }
-    }
-
-    if (typeof item.object !== 'string' && typeof item.object.attributedTo === 'string') {
-        const actor = await lookupActor(apCtx, item.object.attributedTo);
-
-        if (actor) {
-            const json = await actor.toJsonLd();
-            if (typeof json === 'object' && json !== null) {
-                item.object.attributedTo = json;
-            }
-        }
-    }
-
-    // If the object associated with the item is an object with a content property,
-    // we should sanitize the content to prevent XSS (in case it contains HTML)
-    if (item.object && typeof item.object !== 'string' && item.object.content) {
-        item.object.content = sanitizeHtml(item.object.content, {
-            allowedTags: ['a', 'p', 'img', 'br', 'strong', 'em', 'span'],
-            allowedAttributes: {
-                a: ['href'],
-                img: ['src'],
-            }
-        });
-    }
-
-    // If the associated object is a Like, we should check if it's in the provided
-    // liked list and add a liked property to the item if it is
-    let objectId: string = '';
-
-    if (typeof item.object === 'string') {
-        objectId = item.object;
-    } else if (typeof item.object.id === 'string') {
-        objectId = item.object.id;
-    }
-
-    if (objectId) {
-        const likeId = apCtx.getObjectUri(Like, {
-            id: createHash('sha256').update(objectId).digest('hex'),
-        });
-        if (liked.includes(likeId.href)) {
-            if (typeof item.object !== 'string') {
-                item.object.liked = true;
-            }
-        }
-    }
-
-    // If a replies map has been provided, the item is not a string, and the
-    // item has an id, we should nest any replies recursively (which involves
-    // calling this function again for each reply)
-    if (repliesMap && typeof item.object !== 'string' && item.object.id) {
-        item.object.replies = [];
-
-        const replies = repliesMap.get(item.object.id);
-
-        if (replies) {
-            const builtReplies = [];
-
-            for (const reply of replies) {
-                const builtReply = await buildActivity(reply.id, db, apCtx, liked, repliesMap);
-
-                if (builtReply) {
-                    builtReplies.push(builtReply);
-                }
-            }
-
-            item.object.replies = builtReplies;
-        }
-    }
-
-    // Expand the inReplyTo object if it is a string and we are expanding inReplyTo
-    if (expandInReplyTo && typeof item.object !== 'string' && item.object.inReplyTo) {
-        const replyObject = await db.get([item.object.inReplyTo]);
-
-        if (replyObject) {
-            item.object.inReplyTo = replyObject;
-        }
-    }
-
-    // Return the built item
-    return item;
-}
-
 export async function inboxHandler(
     ctx: Context<{ Variables: HonoContextVariables }>,
 ) {
@@ -647,203 +521,4 @@ export async function inboxHandler(
             status: 200,
         },
     );
-}
-
-export async function getActivities(
-    ctx: Context<{ Variables: HonoContextVariables }>,
-) {
-    const DEFAULT_LIMIT = 10;
-
-    const db = ctx.get('db');
-    const globaldb = ctx.get('globaldb');
-    const apCtx = fedify.createContext(ctx.req.raw as Request, {db, globaldb});
-
-    // -------------------------------------------------------------------------
-    // Process query parameters
-    // -------------------------------------------------------------------------
-
-    // Parse "cursor" and "limit" from query parameters
-    // These are used to paginate the results
-    // ?cursor=<string>
-    // ?limit=<number>
-    const queryCursor = ctx.req.query('cursor')
-    const cursor = queryCursor ? Buffer.from(queryCursor, 'base64url').toString('utf-8') : null;
-    const limit = Number.parseInt(ctx.req.query('limit') || DEFAULT_LIMIT.toString(), 10);
-
-    // Parse "includeOwn" from query parameters
-    // This is used to include the user's own activities in the results
-    // ?includeOwn=<boolean>
-    const includeOwn = ctx.req.query('includeOwn') === 'true';
-
-    // Parse "includeReplies" from query parameters
-    // This is used to include nested replies in the results
-    // ?includeReplies=<boolean>
-    const includeReplies = ctx.req.query('includeReplies') === 'true';
-
-    // Parse "filter" from query parameters
-    // This is used to filter the activities by various criteria
-    // ?filter={type: ['<activityType>', '<activityType>:<objectType>', '<activityType>:<objectType>:<criteria>']}
-    const queryFilters = ctx.req.query('filter') || '[]';
-    const filters = JSON.parse(decodeURI(queryFilters))
-
-    const typeFilters = (filters.type || []).map((filter: string) => {
-        const [activityType, objectType = null, criteria = null] = filter.split(':');
-
-        return {
-            activity: activityType,
-            object: objectType,
-            criteria,
-        }
-    });
-
-    console.log('Request query =', ctx.req.query());
-    console.log('Processed query params =', JSON.stringify({
-        cursor,
-        limit,
-        includeOwn,
-        typeFilters,
-    }, null, 2));
-
-    // -------------------------------------------------------------------------
-    // Fetch required data from the database
-    // -------------------------------------------------------------------------
-
-    // Fetch the liked object refs from the database:
-    //   - Data is structured as an array of strings
-    //   - Each string is a URI to an object in the database
-    // This is used to add a "liked" property to the item if the user has liked it
-    const likedRefs = (await db.get<string[]>(['liked'])) || [];
-
-    // Fetch the refs of the activities in the inbox from the database:
-    //   - Data is structured as an array of strings
-    //   - Each string is a URI to an object in the database
-    //   - First item is the oldest, last item is the newest
-    const inboxRefs = ((await db.get<string[]>(['inbox'])) || [])
-
-    // Fetch the refs of the activities in the outbox from the database (if
-    // user is requesting their own activities):
-    //   - Data is structured as an array of strings
-    //   - Each string is a URI to an object in the database
-    //   - First item is the oldest, last item is the newest
-    let outboxRefs: string[] = [];
-
-    if (includeOwn) {
-        outboxRefs = await db.get<string[]>(['outbox']) || [];
-    }
-
-    // To be able to return a sorted / filtered list of activities, we need to
-    // fetch some additional meta data about the referenced activities. Doing this
-    // upfront allows us to sort, filter and paginate the activities before
-    // building them for the response which saves us from having to perform
-    // unnecessary database lookups for referenced activities that will not be
-    // included in the response
-    let activityRefs = [...inboxRefs, ...outboxRefs];
-    const activityMeta = await getActivityMeta(activityRefs);
-
-    // If we can't find the meta data in the database for an activity, we skip
-    // it as this is unexpected
-    activityRefs = activityRefs.filter(ref => activityMeta.has(ref));
-
-    // -------------------------------------------------------------------------
-    // Apply filtering and sorting
-    // -------------------------------------------------------------------------
-
-    // Filter the activity refs by any provided type filters
-    if (typeFilters.length > 0) {
-        activityRefs = activityRefs.filter(ref => {
-            const activity = activityMeta.get(ref)!;
-
-            return typeFilters.some((filter: { activity: string; object: string | null, criteria: string | null }) => {
-                // ?filter={type: ['<activityType>']}
-                if (filter.activity && activity.activity_type !== filter.activity) {
-                    return false;
-                }
-
-                // ?filter={type: ['<activityType>:<objectType>']}
-                if (filter.object && activity.object_type !== filter.object) {
-                    return false;
-                }
-
-                // ?filter={type: ['<activityType>:<objectType>:isReplyToOwn,<siteHost>']}
-                if (filter.criteria && filter.criteria.startsWith('isReplyToOwn,')) {
-                    // If the activity does not have a reply object url or name,
-                    // we can't determine if it's a reply to an own object so
-                    // we skip it
-                    if (!activity.reply_object_url || !activity.reply_object_name) {
-                        return false;
-                    }
-
-                    // Verify that the reply is to an object created by the user by
-                    // checking that the hostname associated with the reply object
-                    // is the same as the hostname of the site. This is not a bullet
-                    // proof check, but it's a good enough for now (i think 😅)
-                    const [_, siteHost] = filter.criteria.split(',');
-                    const { hostname: replyHost } = new URL(activity.reply_object_url);
-
-                    return siteHost === replyHost;
-                }
-
-                return true;
-            });
-        });
-    }
-
-    // Sort the activity refs by the id of the activity (newest first).
-    // We are using the id to sort because currently not all activity types have
-    // a timestamp. The id property is a unique auto incremented number at the
-    // database level
-    activityRefs.sort((a, b) => {
-        return activityMeta.get(b)!.id - activityMeta.get(a)!.id;
-    });
-
-    // -------------------------------------------------------------------------
-    // Paginate
-    // -------------------------------------------------------------------------
-
-    // Find the starting index based on the cursor
-    const startIndex = cursor ? activityRefs.findIndex(ref => ref === cursor) + 1 : 0;
-
-    // Slice the results array based on the cursor and limit
-    const paginatedRefs = activityRefs.slice(startIndex, startIndex + limit);
-
-    // Determine the next cursor
-    const nextCursor = startIndex + paginatedRefs.length < activityRefs.length
-        ? Buffer.from(paginatedRefs[paginatedRefs.length - 1]).toString('base64url')
-        : null;
-
-    // -------------------------------------------------------------------------
-    // Build the activities and return the response
-    // -------------------------------------------------------------------------
-
-    const activities = [];
-
-    // If we need to include replies, fetch the replies map based on the paginated
-    // activity refs, which will be utilised when building the activities
-    const repliesMap = includeReplies
-        ? await getRepliesMap(paginatedRefs)
-        : null;
-
-    // Build the activities
-    for (const ref of paginatedRefs) {
-        try {
-            const builtActivity = await buildActivity(ref, globaldb, apCtx, likedRefs, repliesMap, true);
-
-            if (builtActivity) {
-                activities.push(builtActivity);
-            }
-        } catch (err) {
-            console.log(err);
-        }
-    }
-
-    // Return the response
-    return new Response(JSON.stringify({
-        items: activities,
-        nextCursor,
-    }), {
-        headers: {
-            'Content-Type': 'application/json',
-        },
-        status: 200,
-    });
 }
