+++ src/helpers/activitypub/activity.unit.test.ts
@@ -1,0 +1,3 @@
+import { describe, expect, it, vi } from 'vitest';
+
+describe.skip('buildActivity', () => {});
